chore: fix dts bug
