+++ packages/blade/package.json
@@ -71,7 +71,7 @@
     "clearCache": "jest --clearCache",
     "build:clean": "rm -rf build",
     "build:generate-types": "run-s types:generate-types:*",
-    "build:copy-declarations": "run-s types:copy-declarations:*",
+    "abuild:copy-declarations": "run-s types:copy-declarations:*",
     "types:generate-types:web": "tsc -p ./tsconfig-generate-types.web.json",
     "types:generate-types:native": "tsc -p ./tsconfig-generate-types.native.json",
     "types:typecheck:web": "tsc -p ./tsconfig-typecheck.web.json",
@@ -82,8 +82,8 @@
     "build:clean-declarations": "rm -rf build/types",
     "build:react": "FRAMEWORK=REACT rollup -c && node ./scripts/generateRootImports.js",
     "build:react-native": "FRAMEWORK=REACT_NATIVE rollup -c && node ./scripts/generateRootImports.js",
-    "build:generate-css-variables": "GENERATE_CSS_VARIABLES=true FRAMEWORK=REACT rollup -c && node ./scripts/generateCssThemeVariables.js",
-    "build:clean-theme-bundle": "rm -rf build/js-bundle-for-css",
+    "abuild:generate-css-variables": "GENERATE_CSS_VARIABLES=true FRAMEWORK=REACT rollup -c && node ./scripts/generateCssThemeVariables.js",
+    "abuild:clean-theme-bundle": "rm -rf build/js-bundle-for-css",
     "react-native:get-stories": "sb-rn-get-stories --config-path=./.storybook/react-native && yarn prettier --write ./.storybook/react-native/storybook.requires.js",
     "react-native:storybook:android": "yarn react-native:get-stories && FRAMEWORK=REACT_NATIVE react-native run-android",
     "react-native:storybook:ios": "yarn react-native:get-stories && FRAMEWORK=REACT_NATIVE react-native run-ios",
