+++ packages/blade/src/components/Skeleton/index.ts
@@ -1,0 +1,2 @@
+export type { SkeletonProps } from './types';
+export * from './Skeleton';
