+++ packages/blade/src/components/index.ts
@@ -26,6 +26,7 @@ export * from './Radio';
 export * from './Spinner';
 export * from './SkipNav';
 export * from './Switch';
+export * from './Skeleton';
 export * from './Typography';
 export * from './VisuallyHidden';
 export * from './Amount';
